+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Allow user to select "No icon" when ACF icon select field is set to "Allow null"
 - Add custom settings functionality
+- Increase body padding for the HS widget not to override paging controls
 
 ## [2.8.0] - 2021-05-18
 ### Added